+++ packages/calcite-components/src/components/color-picker/color-picker.tsx
@@ -48,6 +48,7 @@ import {
   RGB_LIMITS,
   SCOPE_SIZE,
   STATIC_DIMENSIONS,
+  ICONS,
 } from "./resources";
 import { Channels, ColorMode, ColorValue, HSLA, HSVA, InternalColor, RGBA } from "./interfaces";
 import T9nStrings from "./assets/t9n/messages.en.json";
@@ -1629,7 +1630,7 @@ export class ColorPicker extends LitElement implements InteractiveComponent {
                     appearance="transparent"
                     class={CSS.deleteColor}
                     disabled={noColor}
-                    iconStart="minus"
+                    iconStart={ICONS.minus}
                     kind="neutral"
                     label={messages.deleteColor}
                     onClick={this.deleteColor}
@@ -1640,7 +1641,7 @@ export class ColorPicker extends LitElement implements InteractiveComponent {
                     appearance="transparent"
                     class={CSS.saveColor}
                     disabled={noColor}
-                    iconStart="plus"
+                    iconStart={ICONS.plus}
                     kind="neutral"
                     label={messages.saveColor}
                     onClick={this.saveColor}