+++ packages/calcite-components/src/components/combobox/combobox.tsx
@@ -61,7 +61,7 @@ import { useSetFocus } from "../../controllers/useSetFocus";
 import { useCancelable } from "../../controllers/useCancelable";
 import T9nStrings from "./assets/t9n/messages.en.json";
 import { ComboboxChildElement, GroupData, ItemData, SelectionDisplay } from "./interfaces";
-import { ComboboxItemGroupSelector, ComboboxItemSelector, CSS, IDS } from "./resources";
+import { ComboboxItemGroupSelector, ComboboxItemSelector, CSS, IDS, ICONS } from "./resources";
 import {
   getItemAncestors,
   getItemChildren,
@@ -76,12 +76,6 @@ declare global {
     "calcite-combobox": Combobox;
   }
 }
-
-const itemUidPrefix = "combobox-item-";
-const chipUidPrefix = "combobox-chip-";
-const labelUidPrefix = "combobox-label-";
-const listboxUidPrefix = "combobox-listbox-";
-const inputUidPrefix = "combobox-input-";
 
 /** @slot - A slot for adding `calcite-combobox-item`s. */
 export class Combobox
@@ -728,7 +722,7 @@ export class Combobox
 
     if (this.allSelected) {
       this.selectedItems.forEach((item) => {
-        const chipEl = this.referenceEl.querySelector<Chip["el"]>(`#${chipUidPrefix}${item.guid}`);
+        const chipEl = this.referenceEl.querySelector<Chip["el"]>(`#${IDS.chip(item.guid)}`);
         if (chipEl) {
           this.hideChip(chipEl);
         }
@@ -1109,7 +1103,7 @@ export class Combobox
 
     if (this.allSelected && this.selectAllEnabled) {
       this.selectedItems.forEach((item) => {
-        const chipEl = this.referenceEl.querySelector<Chip["el"]>(`#${chipUidPrefix}${item.guid}`);
+        const chipEl = this.referenceEl.querySelector<Chip["el"]>(`#${IDS.chip(item.guid)}`);
         if (chipEl) {
           this.hideChip(chipEl);
         }
@@ -1118,7 +1112,7 @@ export class Combobox
 
     if (this.indeterminate) {
       this.selectedItems.forEach((item) => {
-        const chipEl = this.referenceEl.querySelector<Chip["el"]>(`#${chipUidPrefix}${item.guid}`);
+        const chipEl = this.referenceEl.querySelector<Chip["el"]>(`#${IDS.chip(item.guid)}`);
         if (chipEl) {
           this.showChip(chipEl);
         }
@@ -1449,9 +1443,7 @@ export class Combobox
   private focusChip(): void {
     const guid = this.selectedItems[this.activeChipIndex]?.guid;
 
-    const chip = guid
-      ? this.referenceEl.querySelector<Chip["el"]>(`#${chipUidPrefix}${guid}`)
-      : null;
+    const chip = guid ? this.referenceEl.querySelector<Chip["el"]>(`#${IDS.chip(guid)}`) : null;
     chip?.setFocus();
   }
 
@@ -1492,7 +1484,7 @@ export class Combobox
     this.keyboardNavItems.forEach((el, i) => {
       if (i === index) {
         el.active = true;
-        activeDescendant = `${itemUidPrefix}${el.guid}`;
+        activeDescendant = `${IDS.item(el.guid)}`;
       } else {
         el.active = false;
       }
@@ -1562,7 +1554,7 @@ export class Combobox
           data-test-id={`chip-${i}`}
           icon={item.icon}
           iconFlipRtl={item.iconFlipRtl}
-          id={item.guid ? `${chipUidPrefix}${item.guid}` : null}
+          id={item.guid ? `${IDS.chip(item.guid)}` : null}
           key={itemLabel}
           label={label}
           messageOverrides={{ dismissLabel: messages.removeTag }}
@@ -1592,7 +1584,7 @@ export class Combobox
     return (
       <calcite-chip
         class={{
-          chip: true,
+          [CSS.chip]: true,
           [CSS.chipInvisible]: !(this.allSelected && !selectedVisibleChipsCount),
           [CSS.allSelected]: true,
         }}
@@ -1648,7 +1640,7 @@ export class Combobox
     return (
       <calcite-chip
         class={{
-          chip: true,
+          [CSS.chip]: true,
           [CSS.chipInvisible]: chipInvisible,
         }}
         label={label}
@@ -1690,7 +1682,7 @@ export class Combobox
     return (
       <calcite-chip
         class={{
-          chip: true,
+          [CSS.chip]: true,
           [CSS.chipInvisible]: chipInvisible,
         }}
         label={label}
@@ -1712,8 +1704,8 @@ export class Combobox
     return (
       <span
         class={{
-          "input-wrap": true,
-          "input-wrap--single": single,
+          [CSS.inputWrap]: true,
+          [CSS.inputWrapSingle]: single,
         }}
       >
         {showLabel && (
@@ -1729,9 +1721,9 @@ export class Combobox
         )}
         <input
           aria-activedescendant={this.activeDescendant}
-          aria-controls={`${listboxUidPrefix}${guid}`}
+          aria-controls={`${IDS.listbox(guid)}`}
           aria-errormessage={IDS.validationMessage}
-          aria-owns={`${listboxUidPrefix}${guid}`}
+          aria-owns={`${IDS.listbox(guid)}`}
           ariaAutoComplete="list"
           ariaExpanded={open}
           ariaHasPopup="listbox"
@@ -1739,13 +1731,13 @@ export class Combobox
           ariaLabel={getLabelText(this)}
           class={{
             [CSS.input]: true,
-            "input--single": true,
+            [CSS.inputSingle]: true,
             [CSS.inputHidden]: showLabel,
-            "input--icon": this.showingInlineIcon && !!this.placeholderIcon,
+            [CSS.inputIcon]: this.showingInlineIcon && !!this.placeholderIcon,
           }}
           data-test-id="input"
           disabled={disabled}
-          id={`${inputUidPrefix}${guid}`}
+          id={`${IDS.input(guid)}`}
           key="input"
           onFocus={this.comboboxFocusHandler}
           onInput={this.inputHandler}
@@ -1766,7 +1758,7 @@ export class Combobox
       this.createScreenReaderItem({
         ariaLabel: item.label,
         ariaSelected: item.selected,
-        id: `${itemUidPrefix}${item.guid}`,
+        id: `${IDS.item(item.guid)}`,
         textContent: item.heading || item.textLabel,
       }),
     );
@@ -1801,7 +1793,7 @@ export class Combobox
     return (
       <div ariaHidden="true" class={CSS.floatingUIContainer} ref={setFloatingEl}>
         <div class={classes} ref={setContainerEl}>
-          <ul class={{ list: true, "list--hide": !open }}>
+          <ul class={{ [CSS.list]: true, [CSS.listHide]: !open }}>
             {this.selectAllEnabled &&
               this.selectionMode !== "single" &&
               this.selectionMode !== "single-persist" && (
@@ -1833,7 +1825,7 @@ export class Combobox
 
     return (
       this.showingInlineIcon && (
-        <span class="icon-start" key="selected-placeholder-icon">
+        <span class={CSS.iconStart} key="selected-placeholder-icon">
           <calcite-icon
             class={{
               [CSS.selectedIcon]: !showPlaceholder,
@@ -1851,10 +1843,10 @@ export class Combobox
   private renderChevronIcon(): JsxNode {
     const { open } = this;
     return (
-      <span class="icon-end" key="chevron">
+      <span class={CSS.iconEnd} key="chevron">
         <calcite-icon
           class={CSS.icon}
-          icon={open ? "chevron-up" : "chevron-down"}
+          icon={open ? ICONS.chevronUp : ICONS.chevronDown}
           scale={getIconScale(this.scale)}
         />
       </span>
@@ -1885,7 +1877,7 @@ export class Combobox
           {this.renderSelectedOrPlaceholderIcon()}
           <div
             class={{
-              "grid-input": true,
+              [CSS.gridInput]: true,
               [CSS.selectionDisplayFit]: fitSelectionDisplay,
               [CSS.selectionDisplaySingle]: singleSelectionDisplay,
             }}
@@ -1905,8 +1897,8 @@ export class Combobox
               ]}
             <label
               class={CSS.screenReadersOnly}
-              htmlFor={`${inputUidPrefix}${guid}`}
-              id={`${labelUidPrefix}${guid}`}
+              htmlFor={`${IDS.input(guid)}`}
+              id={`${IDS.label(guid)}`}
             >
               {label}
             </label>
@@ -1924,10 +1916,10 @@ export class Combobox
           {!readOnly && this.renderChevronIcon()}
         </div>
         <ul
-          aria-labelledby={`${labelUidPrefix}${guid}`}
+          aria-labelledby={`${IDS.label(guid)}`}
           ariaMultiSelectable="true"
           class={CSS.screenReadersOnly}
-          id={`${listboxUidPrefix}${guid}`}
+          id={`${IDS.listbox(guid)}`}
           role="listbox"
           tabIndex={-1}
         >