+++ packages/calcite-components/src/components/date-picker-day/date-picker-day.tsx
@@ -21,6 +21,7 @@ import { Scale } from "../interfaces";
 import type { DatePicker } from "../date-picker/date-picker";
 import { useSetFocus } from "../../controllers/useSetFocus";
 import { styles } from "./date-picker-day.scss";
+import { CSS } from "./resources";
 
 declare global {
   interface DeclareElements {
@@ -207,9 +208,9 @@ export class DatePickerDay extends LitElement implements InteractiveComponent {
 
     return (
       <InteractiveContainer disabled={this.disabled}>
-        <div ariaHidden="true" class="day-wrapper">
-          <span class="day">
-            <span class="text">{formattedDay}</span>
+        <div ariaHidden="true" class={CSS.dayWrapper}>
+          <span class={CSS.day}>
+            <span class={CSS.text}>{formattedDay}</span>
           </span>
         </div>
       </InteractiveContainer>