+++ packages/calcite-components/src/components/date-picker/date-picker.tsx
@@ -27,7 +27,7 @@ import { HeadingLevel } from "../functional/Heading";
 import { useT9n } from "../../controllers/useT9n";
 import { useSetFocus } from "../../controllers/useSetFocus";
 import T9nStrings from "./assets/t9n/messages.en.json";
-import { DATE_PICKER_FORMAT_OPTIONS, HEADING_LEVEL } from "./resources";
+import { DATE_PICKER_FORMAT_OPTIONS, HEADING_LEVEL, CSS } from "./resources";
 import { DateLocaleData, getLocaleData, getValueAsDateRange } from "./utils";
 import { styles } from "./date-picker.scss";
 
@@ -627,7 +627,7 @@ export class DatePicker extends LitElement {
 
     return (
       <>
-        <div ariaHidden={true} class="container" tabIndex={-1}>
+        <div ariaHidden={true} class={CSS.container} tabIndex={-1}>
           {this.renderMonth(startCalendarActiveDate, this.maxAsDate, minDate, date, endDate)}
         </div>
       </>