+++ packages/calcite-components/src/components/input-date-picker/input-date-picker.tsx
@@ -71,7 +71,7 @@ import type { Label } from "../label/label";
 import type { Input } from "../input/input";
 import { useSetFocus } from "../../controllers/useSetFocus";
 import { styles } from "./input-date-picker.scss";
-import { CSS, IDS } from "./resources";
+import { CSS, ICONS, IDS, POSITION } from "./resources";
 import T9nStrings from "./assets/t9n/messages.en.json";
 import { isTwoDigitYear, normalizeToCurrentCentury } from "./utils";
 
@@ -110,7 +110,7 @@ export class InputDatePicker
 
   defaultValue: InputDatePicker["value"];
 
-  private dialogId = `date-picker-dialog--${guid()}`;
+  private dialogId = IDS.dialog(guid());
 
   private endInput: InputText["el"];
 
@@ -148,7 +148,7 @@ export class InputDatePicker
 
   transitionProp = "opacity" as const;
 
-  private placeholderTextId = `calcite-input-date-picker-placeholder-${guid()}`;
+  private placeholderTextId = IDS.placeholder(guid());
 
   private rangeStartValueChangedByUser = false;
 
@@ -1085,7 +1085,7 @@ export class InputDatePicker
           <div aria-label={this.label} class={CSS.inputContainer} role="group">
             <div
               class={CSS.inputWrapper}
-              data-position="start"
+              data-position={POSITION.start}
               onClick={this.onInputWrapperClick}
               onPointerDown={this.onInputWrapperPointerDown}
               ref={this.setStartWrapper}
@@ -1104,7 +1104,7 @@ export class InputDatePicker
                   [CSS.inputNoRightBorder]: this.range,
                 }}
                 disabled={disabled}
-                icon="calendar"
+                icon={ICONS.calendar}
                 label={this.range ? this.messages.startDate : this.messages.date}
                 oncalciteInputTextInput={this.calciteInternalInputInputHandler}
                 oncalciteInternalInputTextBlur={this.calciteInternalInputBlurHandler}
@@ -1172,7 +1172,7 @@ export class InputDatePicker
             {this.range && (
               <div
                 class={CSS.inputWrapper}
-                data-position="end"
+                data-position={POSITION.end}
                 onClick={this.onInputWrapperClick}
                 onPointerDown={this.onInputWrapperPointerDown}
                 ref={this.setEndWrapper}
@@ -1189,7 +1189,7 @@ export class InputDatePicker
                     [CSS.inputNoRightBorder]: this.layout === "vertical" && this.range,
                   }}
                   disabled={disabled}
-                  icon="calendar"
+                  icon={ICONS.calendar}
                   label={this.messages.endDate}
                   oncalciteInputTextInput={this.calciteInternalInputInputHandler}
                   oncalciteInternalInputTextBlur={this.calciteInternalInputBlurHandler}
@@ -1208,7 +1208,7 @@ export class InputDatePicker
           {this.range && this.layout === "vertical" && (
             <div class={CSS.verticalChevronContainer}>
               <calcite-icon
-                icon={this.open ? "chevron-up" : "chevron-down"}
+                icon={this.open ? ICONS.chevronUp : ICONS.chevronDown}
                 scale={getIconScale(this.scale)}
               />
             </div>
@@ -1234,7 +1234,7 @@ export class InputDatePicker
       <span class={CSS.toggleIcon} tabIndex={-1}>
         <calcite-icon
           class={CSS.chevronIcon}
-          icon={open ? "chevron-up" : "chevron-down"}
+          icon={open ? ICONS.chevronUp : ICONS.chevronDown}
           scale={getIconScale(this.scale)}
         />
       </span>