+++ packages/calcite-components/src/components/input-time-picker/input-time-picker.tsx
@@ -42,7 +42,7 @@ import { useSetFocus } from "../../controllers/useSetFocus";
 import { TimeComponent, useTime } from "../../controllers/useTime";
 import { styles } from "./input-time-picker.scss";
 import T9nStrings from "./assets/t9n/messages.en.json";
-import { CSS, IDS } from "./resources";
+import { CSS, IDS, ICONS } from "./resources";
 
 declare global {
   interface DeclareElements {
@@ -575,7 +575,11 @@ export class InputTimePicker
           ref={this.setContainerEl}
           role="combobox"
         >
-          <calcite-icon class={CSS.clockIcon} icon="clock" scale={scale === "l" ? "m" : "s"} />
+          <calcite-icon
+            class={CSS.clockIcon}
+            icon={ICONS.clock}
+            scale={scale === "l" ? "m" : "s"}
+          />
           <div
             aria-label={getLabelText(this)}
             class={CSS.inputContainer}
@@ -749,7 +753,7 @@ export class InputTimePicker
     return (
       <span class={CSS.toggleIcon} onClick={this.toggleIconClickHandler}>
         <calcite-icon
-          icon={open ? "chevron-up" : "chevron-down"}
+          icon={open ? ICONS.chevronUp : ICONS.chevronDown}
           scale={getIconScale(this.scale)}
         />
       </span>