+++ packages/calcite-components/src/components/rating/resources.ts
@@ -7,8 +7,14 @@ export const CSS = {
   fraction: "fraction",
   numberAverage: "number--average",
   numberCount: "number--count",
+  wrapper: "wrapper",
+  visuallyHidden: "visually-hidden",
+  partial: "partial",
 };
+
+const idPrefix = "calcite-ratings";
 
 export const IDS = {
   validationMessage: "validationMessage",
+  host: (id: any) => `${idPrefix}-${id}` as const,
 };