+++ packages/calcite-components/src/components/sheet/sheet.tsx
@@ -24,7 +24,7 @@ import { usePreventDocumentScroll } from "../../controllers/usePreventDocumentSc
 import { FocusTrapOptions, useFocusTrap } from "../../controllers/useFocusTrap";
 import { resizeStep, resizeShiftStep } from "../../utils/resources";
 import { useSetFocus } from "../../controllers/useSetFocus";
-import { CSS } from "./resources";
+import { CSS, ICONS, IDS } from "./resources";
 import { DisplayMode, ResizeValues } from "./interfaces";
 import T9nStrings from "./assets/t9n/messages.en.json";
 import { styles } from "./sheet.scss";
@@ -328,8 +328,8 @@ export class Sheet extends LitElement implements OpenCloseComponent {
     const { position } = this;
 
     return position === "block-start" || position === "block-end"
-      ? "drag-resize-vertical"
-      : "drag-resize-horizontal";
+      ? ICONS.dragVertical
+      : ICONS.dragHorizontal;
   }
 
   private getContentElDOMRect(): DOMRect {
@@ -613,7 +613,7 @@ export class Sheet extends LitElement implements OpenCloseComponent {
         ref={this.setTransitionEl}
       >
         <calcite-scrim class={CSS.scrim} onClick={this.handleOutsideClose} />
-        <div class={CSS.content} id="sheet-content" ref={this.setContentEl}>
+        <div class={CSS.content} id={IDS.sheetContent} ref={this.setContentEl}>
           <div class={CSS.contentContainer}>
             <slot />
           </div>