+++ packages/calcite-components/src/components/sort-handle/sort-handle.tsx
@@ -17,7 +17,7 @@ import { useT9n } from "../../controllers/useT9n";
 import type { Dropdown } from "../dropdown/dropdown";
 import { useSetFocus } from "../../controllers/useSetFocus";
 import T9nStrings from "./assets/t9n/messages.en.json";
-import { CSS, ICONS, REORDER_VALUES, SUBSTITUTIONS } from "./resources";
+import { CSS, ICONS, REORDER_VALUES, SLOTS, SUBSTITUTIONS } from "./resources";
 import { MoveEventDetail, MoveTo, Reorder, ReorderEventDetail } from "./interfaces";
 import { styles } from "./sort-handle.scss";
 
@@ -277,7 +277,7 @@ export class SortHandle extends LitElement implements InteractiveComponent {
             icon={disabled ? ICONS.blank : ICONS.drag}
             label={text}
             scale={scale}
-            slot="trigger"
+            slot={SLOTS.trigger}
             text={text}
             title={text}
           />