+++ packages/calcite-components/src/components/stepper-item/stepper-item.tsx
@@ -30,7 +30,7 @@ import type { Stepper } from "../stepper/stepper";
 import { isHidden } from "../../utils/component";
 import { useSetFocus } from "../../controllers/useSetFocus";
 import { slotChangeHasContent } from "../../utils/dom";
-import { CSS } from "./resources";
+import { CSS, ICONS } from "./resources";
 import T9nStrings from "./assets/t9n/messages.en.json";
 import { styles } from "./stepper-item.scss";
 
@@ -372,18 +372,18 @@ export class StepperItem extends LitElement implements InteractiveComponent {
   }
 
   private renderIcon(): JsxNode {
-    let path: IconNameOrString = "circle";
+    let path: IconNameOrString = ICONS.circle;
 
     if (this.selected && (this.layout !== "horizontal-single" || (!this.error && !this.complete))) {
-      path = "circleF";
+      path = ICONS.circleF;
     } else if (this.error) {
-      path = "exclamationMarkCircleF";
+      path = ICONS.exclamationMarkCircleF;
     } else if (this.complete) {
-      path = "checkCircleF";
+      path = ICONS.checkCircleF;
     }
 
     return (
-      <calcite-icon class="stepper-item-icon" flipRtl={this.iconFlipRtl} icon={path} scale="s" />
+      <calcite-icon class={CSS.stepperItemIcon} flipRtl={this.iconFlipRtl} icon={path} scale="s" />
     );
   }