+++ packages/calcite-components/src/components/time-picker/time-picker.tsx
@@ -12,7 +12,7 @@ import { getElementDir } from "../../utils/dom";
 import { useT9n } from "../../controllers/useT9n";
 import { useSetFocus } from "../../controllers/useSetFocus";
 import { TimeComponent, useTime } from "../../controllers/useTime";
-import { CSS } from "./resources";
+import { CSS, ICONS } from "./resources";
 import T9nStrings from "./assets/t9n/messages.en.json";
 import { styles } from "./time-picker.scss";
 
@@ -403,7 +403,7 @@ export class TimePicker extends LitElement implements TimeComponent {
           [CSS.timePicker]: true,
           [CSS.showMeridiem]: showMeridiem,
           [CSS.showSecond]: this.showSecond,
-          [CSS[`scale-${scale}`]]: true,
+          [CSS.scale(scale)]: true,
         }}
         dir="ltr"
       >
@@ -418,7 +418,7 @@ export class TimePicker extends LitElement implements TimeComponent {
             onClick={this.hourUpClickHandler}
             role="button"
           >
-            <calcite-icon icon="chevron-up" scale={iconScale} />
+            <calcite-icon icon={ICONS.chevronUp} scale={iconScale} />
           </span>
           <span
             ariaLabel={messages.hour}
@@ -450,7 +450,7 @@ export class TimePicker extends LitElement implements TimeComponent {
             onClick={this.hourDownClickHandler}
             role="button"
           >
-            <calcite-icon icon="chevron-down" scale={iconScale} />
+            <calcite-icon icon={ICONS.chevronDown} scale={iconScale} />
           </span>
         </div>
         <span class={{ [CSS.delimiter]: true, [CSS.hourSuffix]: true }}>{localizedHourSuffix}</span>
@@ -464,7 +464,7 @@ export class TimePicker extends LitElement implements TimeComponent {
             onClick={this.minuteUpClickHandler}
             role="button"
           >
-            <calcite-icon icon="chevron-up" scale={iconScale} />
+            <calcite-icon icon={ICONS.chevronUp} scale={iconScale} />
           </span>
           <span
             ariaLabel={messages.minute}
@@ -495,7 +495,7 @@ export class TimePicker extends LitElement implements TimeComponent {
             onClick={this.minuteDownClickHandler}
             role="button"
           >
-            <calcite-icon icon="chevron-down" scale={iconScale} />
+            <calcite-icon icon={ICONS.chevronDown} scale={iconScale} />
           </span>
         </div>
         {this.showSecond && (
@@ -514,7 +514,7 @@ export class TimePicker extends LitElement implements TimeComponent {
               onClick={this.secondUpClickHandler}
               role="button"
             >
-              <calcite-icon icon="chevron-up" scale={iconScale} />
+              <calcite-icon icon={ICONS.chevronUp} scale={iconScale} />
             </span>
             <span
               ariaLabel={messages.second}
@@ -545,7 +545,7 @@ export class TimePicker extends LitElement implements TimeComponent {
               onClick={this.secondDownClickHandler}
               role="button"
             >
-              <calcite-icon icon="chevron-down" scale={iconScale} />
+              <calcite-icon icon={ICONS.chevronDown} scale={iconScale} />
             </span>
           </div>
         )}
@@ -565,7 +565,7 @@ export class TimePicker extends LitElement implements TimeComponent {
               onClick={this.fractionalSecondUpClickHandler}
               role="button"
             >
-              <calcite-icon icon="chevron-up" scale={iconScale} />
+              <calcite-icon icon={ICONS.chevronUp} scale={iconScale} />
             </span>
             <span
               ariaLabel={messages.fractionalSecond}
@@ -596,7 +596,7 @@ export class TimePicker extends LitElement implements TimeComponent {
               onClick={this.fractionalSecondDownClickHandler}
               role="button"
             >
-              <calcite-icon icon="chevron-down" scale={iconScale} />
+              <calcite-icon icon={ICONS.chevronDown} scale={iconScale} />
             </span>
           </div>
         )}
@@ -623,7 +623,7 @@ export class TimePicker extends LitElement implements TimeComponent {
               onClick={this.meridiemUpClickHandler}
               role="button"
             >
-              <calcite-icon icon="chevron-up" scale={iconScale} />
+              <calcite-icon icon={ICONS.chevronUp} scale={iconScale} />
             </span>
             <span
               ariaLabel={messages.meridiem}
@@ -655,7 +655,7 @@ export class TimePicker extends LitElement implements TimeComponent {
               onClick={this.meridiemDownClickHandler}
               role="button"
             >
-              <calcite-icon icon="chevron-down" scale={iconScale} />
+              <calcite-icon icon={ICONS.chevronDown} scale={iconScale} />
             </span>
           </div>
         )}